requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.1.3
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-storages==1.14.4
dropbox==12.0.2
Faker==33.0.0
idna==3.10
ply==3.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
stone==3.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
